Avoid memory pressure when scheduling lots of timeouts
If timeouts are scheduled very frequently which are less than the cutoff
time, then the near list will grow nearly unbounded. When canceling the
timeout, if it wasn't in the far list, then try removing it from the
near list.

This will also have the effect of keeping the near list shorter which
will be better for performance.